requirements.txt
PyQt5
ebooklib
pydub
google-cloud-speech
mutagen
